+++ crates/openshell-cli/src/commands/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod common;
+pub mod gateway;
